Added documentation, and annotation. Fixed some naming issues.
