+++ .gitignore
@@ -1,3 +1,3 @@
 # Created by .ignore support plugin (hsz.mobi)
 .idea
-/app-generation/generated/*
+/project-generation/generated/*
