--- app-generation/content/templates/.gitignore.tmpl
@@ -1,10 +1,0 @@
-# Mac
-.DS_Store
-
-#VSCode
-.vscode
-
-# Intellij
-.idea/
-*.iml
-*.iws
