--- app-generation/content/templates/ci/scripts/build.tmpl
@@ -1,10 +1,0 @@
-#!/bin/bash -eux
-
-cwd=$(pwd)
-
-export GOPATH=$cwd/go
-
-pushd $GOPATH/src/github.com/ONSdigital/{{.Name}}
-  make build && cp build/{{.Name}} $cwd/build
-  cp Dockerfile.concourse $cwd/build
-popd
