--- app-generation/content/templates/ci/scripts/unit.tmpl
@@ -1,7 +1,0 @@
-#!/bin/bash -eux
-
-export GOPATH=$(pwd)/go
-
-pushd $GOPATH/src/github.com/ONSdigital/{{.Name}}
-  make test
-popd
