--- app-generation/content/templates/ci/unit.tmpl
@@ -1,16 +1,0 @@
----
-
-platform: linux
-
-image_resource:
-  type: docker-image
-  source:
-    repository: golang
-    tag: {{.GoVersion}}
-
-inputs:
-  - name: dp-frontend-dataset-controller
-    path: go/src/github.com/ONSdigital/{{.Name}}
-
-run:
-  path: go/src/github.com/ONSdigital/{{.Name}}/ci/scripts/unit.sh
