--- app-generation/content/templates/config/config.tmpl
@@ -1,25 +1,0 @@
-package config
-
-import "github.com/kelseyhightower/envconfig"
-
-// Config represents service configuration for {{.Name}}
-type Config struct {
-	BindAddr             string `envconfig:"BIND_ADDR"`
-}
-
-var cfg *Config
-
-// Get returns the default config with any modifications through environment
-// variables
-func Get() (*Config, error) {
-	if cfg != nil {
-		return cfg, nil
-	}
-
-	cfg := &Config{
-		BindAddr:             ":{{.Port}}",
-
-	}
-
-	return cfg, envconfig.Process("", cfg)
-}
