--- app-generation/content/templates/handlers/handlers_test.tmpl
@@ -1,5 +1,0 @@
-package handlers
-
-import (
-
-)
