--- app-generation/content/templates/mapper/mapper.tmpl
@@ -1,24 +1,0 @@
-package mapper
-
-import (
-	"net/http"
-	"github.com/ONSdigital/go-ns/log"
-)
-
-// ClientError is an interface that can be used to retrieve the status code if a client has errored
-type ClientError interface {
-	Error() string
-	Code() int
-}
-
-func setStatusCode(req *http.Request, w http.ResponseWriter, err error) {
-	status := http.StatusInternalServerError
-	if err, ok := err.(ClientError); ok {
-		if err.Code() == http.StatusNotFound {
-			status = err.Code()
-		}
-	}
-	log.ErrorCtx(req.Context(), err, log.Data{"setting-response-status": status})
-	w.WriteHeader(status)
-}
-
