--- app-generation/content/templates/mapper/mapper_test.tmpl
@@ -1,5 +1,0 @@
-package mapper
-
-import (
-
-)
