--- app-generation/content/templates/routes/routes_test.tmpl
@@ -1,5 +1,0 @@
-package routes
-
-import (
-
-)
