+++ project-generation/content/templates/.github/ISSUES_TEMPLATE.tmpl
@@ -1,0 +1,19 @@
+### Expected behaviour
+
+Describe the issue; with what you expect to happen
+
+### Actual behaviour
+
+Describe the issue; is currently happening
+
+### Steps to reproduce the issue
+
+Provide a detailed step by step method on how to reliably recreate the issue
+
+### Environment information
+
+Provide information relevant to the issue such as 
+* Operating system
+* Browser
+* Relevant environment variables
+* Feature flags if known
