+++ project-generation/content/templates/.github/PULL_REQUEST_TEMPLATE.tmpl
@@ -1,0 +1,11 @@
+### What
+
+Describe what you have changed and why.
+
+### How to review
+
+Describe the steps required to test the changes.
+
+### Who can review
+
+Describe who worked on the changes, so that other people can review.
