+++ project-generation/content/templates/.gitignore.tmpl
@@ -1,0 +1,10 @@
+# Mac
+.DS_Store
+
+#VSCode
+.vscode
+
+# Intellij
+.idea/
+*.iml
+*.iws
