+++ project-generation/content/templates/ci/build.tmpl
@@ -1,0 +1,19 @@
+---
+
+platform: linux
+
+image_resource:
+  type: docker-image
+  source:
+    repository: golang
+    tags: {{.GoVersion}}
+
+inputs:
+  - name: dp-frontend-dataset-controller
+    path: go/src/github.com/ONSdigital/{{.Name}}
+
+outputs:
+  - name: build
+
+run:
+  path: go/src/github.com/ONSdigital/{{.Name}}/ci/scripts/build.sh
