+++ project-generation/content/templates/ci/scripts/unit.tmpl
@@ -1,0 +1,7 @@
+#!/bin/bash -eux
+
+export GOPATH=$(pwd)/go
+
+pushd $GOPATH/src/github.com/ONSdigital/{{.Name}}
+  make test
+popd
