+++ project-generation/content/templates/handlers/handlers_test.tmpl
@@ -1,0 +1,5 @@
+package handlers
+
+import (
+
+)
