+++ project-generation/content/templates/mapper/mapper_test.tmpl
@@ -1,0 +1,5 @@
+package mapper
+
+import (
+
+)
