+++ project-generation/content/templates/readme.tmpl
@@ -1,0 +1,18 @@
+{{.Name}}
+================
+
+### Getting started
+
+#### Dependencies
+
+### Configuration
+
+### Contributing
+
+See [CONTRIBUTING](CONTRIBUTING.md) for details.
+
+### License
+
+Copyright © {{.Year}}, Office for National Statistics (https://www.ons.gov.uk)
+
+Released under MIT license, see [LICENSE](LICENSE.md) for details.
