+++ project-generation/content/templates/routes/routes_test.tmpl
@@ -1,0 +1,5 @@
+package routes
+
+import (
+
+)
