+++ repository-creation/README.md
@@ -1,0 +1,20 @@
+## Repo Generation Tool
+
+### Usage prerequisite 
+
+#### Github 'Personal Access Token'
+    1. Navigate to github.com
+    2. Click your Avatar
+    3. Go to Settings
+    4. Go to Developer Settings
+    5. Click Personal Access Token
+    6. Generate an Access token with permissions to manipulate repositories
+    
+#### Be a member of the ONS Digital Publishing team and the ONSDigital organisation with permissions to create new repositories
+___
+### How to use
+1. Run ./dp-utils repo-generation github
+2. Enter your 'Personal Access Token' when prompted 
+3. Enter your Github handler when prompted
+4. Enter your repository name, note that if the repository is specific to Digital Publishing then prefix it with 'dp-'
+5. Enter a description for the repository when prompted
